- pass token to api - separate authentication configuration into extension method - Explicit usage of roles
