+++ Mango.Services.Coupon.Api/Controllers/CouponAPIController.cs
@@ -2,14 +2,13 @@
 using Mango.Services.Coupon.Api.Data;
 using Mango.Services.Coupon.Api.Models;
 using Mango.Services.Coupon.Api.Models.DTOs;
-using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Mango.Services.Coupon.Api.Controllers
 {
     [Route("api/coupon")]
     [ApiController]
-    [Authorize(Roles = "ADMIN")]
+    //[Authorize(Roles = "ADMIN")]
     public class CouponAPIController : ControllerBase
     {
         private readonly AppDbContext _db;
