+++ Mango.Services.Coupon.Api/Program.cs
@@ -1,11 +1,10 @@
 using AutoMapper;
 using Mango.Services.Coupon.Api;
 using Mango.Services.Coupon.Api.Data;
+using Mango.Services.Coupon.Api.Extensions;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.IdentityModel.Tokens;
 using Microsoft.OpenApi.Models;
-using System.Text;
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -43,29 +42,9 @@ builder.Services.AddSwaggerGen(options =>
         }
     });
 });
-var sectionSettings = builder.Configuration.GetSection("ApiSettings");
-var secret = sectionSettings.GetValue<string>("Secret");
-var issuer = sectionSettings.GetValue<string>("Issuer");
-var audience = sectionSettings.GetValue<string>("Audience");
 
-var key = Encoding.ASCII.GetBytes(secret);
-
-builder.Services.AddAuthentication(x =>
-{
-    x.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
-    x.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
-}).AddJwtBearer(x =>
-{
-    x.TokenValidationParameters = new TokenValidationParameters()
-    {
-        ValidateIssuer = true,
-        ValidateAudience = true,
-        ValidateIssuerSigningKey = true,
-        ValidIssuer = issuer,
-        IssuerSigningKey = new SymmetricSecurityKey(key),
-        ValidAudience = audience,
-    };
-});
+// Our new extension method for adding authentication configuration...
+builder.AddApplicationAuthentication();
 
 builder.Services.AddAuthorization();
