+++ Mango.Web/Services/IServices/IBaseService.cs
@@ -4,6 +4,6 @@ namespace Mango.Web.Services.IServices
 {
     public interface IBaseService
     {
-        Task<ResponseDTO?> SendAsync(RequestDto request);
+        Task<ResponseDTO?> SendAsync(RequestDto request, bool withTokenEnabled = true);
     }
 }
